Completes the assignment
This project was my first experience trying to create something in the game engine Godot. What I created was a rudimentary framework to create a game like brick breaker. Through the process of creating this assignment, I have learned the basics of creating nodes, and how to differentiate the properties of each of these nodes. This is just the framework for creating a game like brick breaker.